Fix changes detection on websql for table with compound keys.
The websql driver split compound index in n columns, when parsing schema
from database, the driver read each column as a separated index. So when
trying to detect changes, the websql always detect that the number of
index on a store with compound index is not the same.
This patch split the compound index on schema before comparaison.