Fix #90 : For index name comparaison indexOf is used instead of ==
When using indexOf instead of == if a index name partially match
compouned index name, both index are considered as equal by the code.
IndexOf seems to be used because 'name' var is threated as an array.
getName definition states that return type is a string so this fix
should be safe.